pyproject.toml
[project]
name = "transfer-file"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "mysql-connector-python>=9.2.0",
]
